requirements.txt
dspy
chess
pyyaml
anthropic
